refactor(validation): 검증 기능 모듈 리팩토링
- 로또 번호 규칙 상수 적용
- checkRuleSet 파일 위치 utils로 이동
